restructure everyting
