+++ src/micropython/audio.py
@@ -1,0 +1,2 @@
+def play():
+    print("audio.play")
